+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - Fixed tokenid crash
 - Removed delay when closing contract details
+
+##### [2.4.0] - 18 September 23
+
+- Tweak WalletSelect
+- Tweak block time check on form